Move web package contents to handlers package
