+++ internal/handlers/update.go
@@ -3,7 +3,6 @@ package handlers
 import (
 	"errors"
 	"github.com/matthiasBT/monitoring/internal/storage"
-	"github.com/matthiasBT/monitoring/internal/web"
 	"net/http"
 )
 
@@ -12,18 +11,18 @@ func UpdateMetric(w http.ResponseWriter, r *http.Request, patternUpdate string, 
 		w.WriteHeader(http.StatusMethodNotAllowed)
 		return
 	}
-	metricUpdate, err := web.ParseMetricUpdate(r.URL.Path, patternUpdate)
+	metricUpdate, err := ParseMetricUpdate(r.URL.Path, patternUpdate)
 	if err == nil {
 		storage.Add(*metricUpdate)
 		w.WriteHeader(http.StatusOK)
 		return
 	}
 	switch {
-	case errors.Is(err, web.ErrInvalidMetricType):
+	case errors.Is(err, ErrInvalidMetricType):
 		http.Error(w, err.Error(), http.StatusBadRequest)
-	case errors.Is(err, web.ErrMissingMetricName):
+	case errors.Is(err, ErrMissingMetricName):
 		http.Error(w, err.Error(), http.StatusNotFound)
-	case errors.Is(err, web.ErrInvalidMetricVal):
+	case errors.Is(err, ErrInvalidMetricVal):
 		http.Error(w, err.Error(), http.StatusBadRequest)
 	}
 }
