+++ internal/handlers/url.go
@@ -1,4 +1,4 @@
-package web
+package handlers
 
 import (
 	"errors"
